Update index.md
Updated helm search to query for nginx and provided a jq command to set the version. As written, the command to install nginx fails because the version is not set. Furthermore, we search helm for postgres, I changed the search criteria to look for nginx instead to be consistent with the rest of the example exercise.